pyexfil/Stega/datamatrix/requirements.txt
pylibdmtx
opencv-python
pillow